PAD-ChatService/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
psycopg2-binary==2.9.7
python-dotenv==1.0.0
eventlet==0.33.3